refactor(shell): co-locate component, merged unit tests, E2E in InteractiveShell/ folder
Moves InteractiveShell.tsx + .module.css into components/client/InteractiveShell/,
adds barrel index.ts, merges streaming and a11y unit tests into one co-located
InteractiveShell.test.tsx, and moves tests/e2e/ask.spec.ts to
InteractiveShell.e2e.ts (picked up by the chromium-components Playwright projects).
Removes ask from the testMatch alternation in playwright.config.ts.
