+++ components/client/InteractiveShell/index.ts
@@ -1,0 +1,1 @@
+export { InteractiveShell } from './InteractiveShell';
